Move GameRenderer to client/renderers folder.
